+++ src/helpers/NetworkHelper.ts
@@ -1,7 +1,7 @@
 import { JsonRpcProvider, StaticJsonRpcProvider } from "@ethersproject/providers";
 
 import { NetworkId, NETWORKS } from "../constants";
-import { Providers } from "./providers/Providers";
+import { Providers } from "./providers/Providers/Providers";
 
 interface IGetCurrentNetwork {
   provider: StaticJsonRpcProvider | JsonRpcProvider;
